chore: move howto download into component
