+++ src/pages/Howto/Content/Howto/HowtoDownloads/downloadCooldown.ts
